src/documentation.py
```python
# Importing necessary libraries
from sphinx.application import Sphinx
from jupyter_client.manager import start_new_kernel
from atlassian import Confluence
from docutils.core import publish_string

class Documentation:
    def __init__(self, model):
        self.model = model

    def sphinx_documentation(self, source_dir, build_dir, config_dir):
        """
        Function to generate documentation using Sphinx
        """
        app = Sphinx(source_dir, config_dir, build_dir, build_dir, 'html')
        app.build()
        return app.statuscode

    def jupyter_notebook_documentation(self, notebook_path):
        """
        Function to generate documentation using Jupyter Notebook
        """
        kernel_manager, kernel_client = start_new_kernel()

        with open(notebook_path) as f:
            nb = nbformat.read(f, as_version=4)

        ep = ExecutePreprocessor(timeout=600, kernel_name='python3')
        ep.preprocess(nb, {'metadata': {'path': 'notebooks/'}})

        with open(notebook_path, 'wt') as f:
            nbformat.write(nb, f)

        return nb

    def confluence_documentation(self, page_id, title, body):
        """
        Function to generate documentation using Confluence
        """
        confluence = Confluence(
            url='https://your-domain.atlassian.net/wiki',
            username='email@example.com',
            password='api_token')

        status = confluence.update_page(
            parent_id = '123456',
            page_id = page_id,
            title = title,
            body = body)

        return status

    def read_the_docs_documentation(self, docstring):
        """
        Function to generate documentation using Read the Docs
        """
        html = publish_string(source=docstring, writer_name='html')
        return html
```
