tests/test_documentation.py
```python
# Importing necessary libraries
import unittest
from src.documentation import Documentation

class TestDocumentation(unittest.TestCase):
    def setUp(self):
        self.documentation = Documentation('test_model')

    def test_sphinx_documentation(self):
        statuscode = self.documentation.sphinx_documentation('source_dir', 'build_dir', 'config_dir')
        self.assertEqual(statuscode, 0)

    def test_jupyter_notebook_documentation(self):
        nb = self.documentation.jupyter_notebook_documentation('notebook_path')
        self.assertIsNotNone(nb)

    def test_confluence_documentation(self):
        status = self.documentation.confluence_documentation('page_id', 'title', 'body')
        self.assertTrue(status)

    def test_read_the_docs_documentation(self):
        html = self.documentation.read_the_docs_documentation('docstring')
        self.assertIsNotNone(html)

if __name__ == '__main__':
    unittest.main()
```
